# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface PasskeySupported {
  supported: boolean;
  message?: string;
}

export interface PasskeyRegisterOptionsBody {
  /**
   * @minLength 1
   * @maxLength 200
   */
  username: string;
}

/**
 * Options object returned by simplewebauthn for navigator.credentials.create
 */
export interface PasskeyRegisterOptionsResponse {
  [key: string]: unknown;
}

export type PasskeyRegisterVerifyBodyResponse = { [key: string]: unknown };

export interface PasskeyRegisterVerifyBody {
  /**
   * @minLength 1
   * @maxLength 200
   */
  username: string;
  response: PasskeyRegisterVerifyBodyResponse;
}

export interface UserPublic {
  id: string;
  username: string;
  email?: string | null;
}

export interface PasskeyRegisterVerifyResponse {
  success: boolean;
  message?: string;
  userId: string;
  user: UserPublic;
  token: string;
  expiresAt: number;
}

export interface PasskeyAuthenticateOptionsBody {
  /**
   * @minLength 1
   * @maxLength 200
   */
  username?: string;
}

/**
 * Options object returned by simplewebauthn for navigator.credentials.get
 */
export interface PasskeyAuthenticateOptionsResponse {
  [key: string]: unknown;
}

export type PasskeyAuthenticateVerifyBodyResponse = { [key: string]: unknown };

export interface PasskeyAuthenticateVerifyBody {
  response: PasskeyAuthenticateVerifyBodyResponse;
}

export interface PasskeyAuthenticateVerifyResponse {
  success: boolean;
  message?: string;
  user: UserPublic;
  token: string;
  expiresAt: number;
}

export interface SessionTokenResponse {
  success: boolean;
  message?: string;
  token: string;
  expiresAt: number;
}

export interface UserMeResponse {
  id?: string;
  username: string;
  displayName: string;
  email?: string | null;
  groups?: string[];
  isAuthenticated: boolean;
}

export type TtsSpeechBodyModel =
  (typeof TtsSpeechBodyModel)[keyof typeof TtsSpeechBodyModel];

export const TtsSpeechBodyModel = {
  "tts-1": "tts-1",
  "tts-1-hd": "tts-1-hd",
  "gpt-4o-mini-tts": "gpt-4o-mini-tts",
} as const;

export interface TtsSpeechBody {
  /** @minLength 1 */
  text: string;
  voice?: string;
  /**
   * @minimum 0.25
   * @maximum 4
   */
  speed?: number;
  model?: TtsSpeechBodyModel;
}

export interface AudioTranscriptionsResponse {
  text: string;
}

export interface TtsHealthResponse {
  status: string;
  storage?: string;
  fetchAvailable?: boolean;
}

export interface ScriptMeta {
  id: string;
  title: string;
  author?: string;
  description: string;
  language: string;
}

export interface ScriptListResponse {
  scripts: ScriptMeta[];
}

export interface ScriptLine {
  speaker: string;
  line: string;
}

export interface ScriptContent {
  lines: ScriptLine[];
  [key: string]: unknown;
}

export interface ScriptResponse {
  meta: ScriptMeta;
  content: ScriptContent;
}

/**
 * multipart form-data with file, optional model and language
 */
export type AudioTranscriptionsBody = { [key: string]: unknown };