# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /audio/transcriptions:
    post:
      operationId: audioTranscriptions
      tags: [audio]
      summary: OpenAI audio transcription proxy
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              description: multipart form-data with file, optional model and language
              additionalProperties: true
      responses:
        "200":
          description: Transcription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AudioTranscriptionsResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  schemas:
...
    AudioTranscriptionsResponse:
      type: object
      properties:
        text: { type: string }
      required: [text]
...