# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Server-side passkey support flag
 */
export const PasskeySupportedResponse = zod.object({
  supported: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary Generate WebAuthn registration options
 */
export const passkeyRegisterOptionsBodyUsernameMax = 200;

export const PasskeyRegisterOptionsBody = zod.object({
  username: zod.string().min(1).max(passkeyRegisterOptionsBodyUsernameMax),
});

export const PasskeyRegisterOptionsResponse = zod
  .record(zod.string(), zod.unknown())
  .describe(
    "Options object returned by simplewebauthn for navigator.credentials.create",
  );

/**
 * @summary Verify WebAuthn registration response
 */
export const passkeyRegisterVerifyBodyUsernameMax = 200;

export const PasskeyRegisterVerifyBody = zod.object({
  username: zod.string().min(1).max(passkeyRegisterVerifyBodyUsernameMax),
  response: zod.record(zod.string(), zod.unknown()),
});

export const PasskeyRegisterVerifyResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
  userId: zod.string(),
  user: zod.object({
    id: zod.string(),
    username: zod.string(),
    email: zod.union([zod.string(), zod.null()]).optional(),
  }),
  token: zod.string(),
  expiresAt: zod.number(),
});

/**
 * @summary Generate WebAuthn authentication options
 */
export const passkeyAuthenticateOptionsBodyUsernameMax = 200;

export const PasskeyAuthenticateOptionsBody = zod.object({
  username: zod
    .string()
    .min(1)
    .max(passkeyAuthenticateOptionsBodyUsernameMax)
    .optional(),
});

export const PasskeyAuthenticateOptionsResponse = zod
  .record(zod.string(), zod.unknown())
  .describe(
    "Options object returned by simplewebauthn for navigator.credentials.get",
  );

/**
 * @summary Verify WebAuthn authentication response
 */
export const PasskeyAuthenticateVerifyBody = zod.object({
  response: zod.record(zod.string(), zod.unknown()),
});

export const PasskeyAuthenticateVerifyResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
  user: zod.object({
    id: zod.string(),
    username: zod.string(),
    email: zod.union([zod.string(), zod.null()]).optional(),
  }),
  token: zod.string(),
  expiresAt: zod.number(),
});

/**
 * @summary Refresh session token
 */
export const PasskeyRefreshResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
  token: zod.string(),
  expiresAt: zod.number(),
});

/**
 * @summary Logout (revoke session token)
 */
export const PasskeyLogoutResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary Logout (revoke session token) — alias of /passkey/logout
 */
export const UserLogoutResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary Current authenticated user
 */
export const GetUserMeResponse = zod.object({
  id: zod.string().optional(),
  username: zod.string(),
  displayName: zod.string(),
  email: zod.union([zod.string(), zod.null()]).optional(),
  groups: zod.array(zod.string()).optional(),
  isAuthenticated: zod.boolean(),
});

/**
 * @summary TTS proxy health
 */
export const TtsHealthResponse = zod.object({
  status: zod.string(),
  storage: zod.string().optional(),
  fetchAvailable: zod.boolean().optional(),
});

/**
 * @summary List built-in scripts
 */
export const ListScriptsResponse = zod.object({
  scripts: zod.array(
    zod.object({
      id: zod.string(),
      title: zod.string(),
      author: zod.string().optional(),
      description: zod.string(),
      language: zod.string(),
    }),
  ),
});

/**
 * @summary Get a script by id
 */
export const GetScriptParams = zod.object({
  id: zod.coerce.string(),
});

export const GetScriptResponse = zod.object({
  meta: zod.object({
    id: zod.string(),
    title: zod.string(),
    author: zod.string().optional(),
    description: zod.string(),
    language: zod.string(),
  }),
  content: zod.object({
    lines: zod.array(
      zod.object({
        speaker: zod.string(),
        line: zod.string(),
      }),
    ),
  }),
});

/**
 * @summary OpenAI text-to-speech proxy
 */

export const ttsSpeechBodySpeedMin = 0.25;
export const ttsSpeechBodySpeedMax = 4;

export const TtsSpeechBody = zod.object({
  text: zod.string().min(1),
  voice: zod.string().optional(),
  speed: zod
    .number()
    .min(ttsSpeechBodySpeedMin)
    .max(ttsSpeechBodySpeedMax)
    .optional(),
  model: zod.enum(["tts-1", "tts-1-hd", "gpt-4o-mini-tts"]).optional(),
});

/**
 * @summary OpenAI audio transcription proxy
 */
export const AudioTranscriptionsBody = zod
  .record(zod.string(), zod.unknown())
  .describe("multipart form-data with file, optional model and language");

export const AudioTranscriptionsResponse = zod.object({
  text: zod.string(),
});